+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Next version
   Django 5.0. Python 3.10 and Django 4.2 are still tested so we should be safe.
 - Added list style overrides to hopefully make lists usable with the Grappelli
   admin skin.
+- Disallowed overriding the default editor preset.
 
 
 0.10 (2024-12-17)